further improve robustness of non-strict mode
this time, we make `NamedItemList` ignore objects that do not feature
a non-empty string `.short_name`. This should fix
https://github.com/mercedes-benz/odxtools/issues/449 .

Note that the code currently still accepts short name strings which
are not valid python identifiers, e.g., `%foo $bar!`. Since these are
not allowed by the ODX standard and (hopefully) do not cause any
exceptions to be raised, fixing this issue is not really urgent IMO...

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>
Approved-by: Michael Hahn <michael.a.hahn@mercedes-benz.com>